AWB Updater Bump to 1.3.5.0
Will handle Wikidiff2.dll if it exists, or Diff.dll if that exists

git-svn-id: https://svn.code.sf.net/p/autowikibrowser/code/AWB@1887 5227b50d-9349-4574-b055-4b0a7525e11c

+++ AWBUpdater/Updater.cs
@@ -312,10 +312,16 @@ AWBUpdater will now close!");
             if (awbUpdate)
             {
                 File.Copy(tempDirectory + "AutoWikiBrowser.exe", AWBdirectory + "AutoWikiBrowser.exe", true);
+                File.Copy(tempDirectory + "AutoWikiBrowser.exe.config", AWBdirectory + "AutoWikiBrowser.exe.config", true);
+
                 File.Copy(tempDirectory + "WikiFunctions.dll", AWBdirectory + "WikiFunctions.dll", true);
                 File.Copy(tempDirectory + "IRCMonitor.exe", AWBdirectory + "IRCMonitor.exe", true);
-                File.Copy(tempDirectory + "Diff.dll", AWBdirectory + "Diff.dll", true);
-                File.Copy(tempDirectory + "AutoWikiBrowser.exe.config", AWBdirectory + "AutoWikiBrowser.exe.config", true);
+                if (File.Exists(tempDirectory + "Diff.dll"))
+                    File.Copy(tempDirectory + "Diff.dll", AWBdirectory + "Diff.dll", true);
+
+                if (File.Exists(tempDirectory + "Wikidiff2.dll"))
+                    File.Copy(tempDirectory + "Wikidiff2.dll", AWBdirectory + "Wikidiff2.dll", true);
+                //File.Delete(AWBdirectory + "Wikidiff2.dll");
 
                 File.Copy(tempDirectory + "gpl-2.0.txt", AWBdirectory + "gpl-2.0.txt", true);
                 File.Copy(tempDirectory + "gpl-3.0.txt", AWBdirectory + "gpl-3.0.txt", true);
